implemented using mini-gmp in Theory.hpp
